+++ r-sql.tf
@@ -30,7 +30,7 @@ resource "azurerm_mssql_server" "sql" {
 }
 
 resource "azurerm_mssql_firewall_rule" "firewall_rule" {
-  count = length(var.allowed_cidr_list)
+  count = try(length(var.allowed_cidr_list), 0)
 
   name      = "rule-${count.index}"
   server_id = azurerm_mssql_server.sql.id
@@ -70,7 +70,7 @@ resource "azurerm_mssql_elasticpool" "elastic_pool" {
 }
 
 resource "azurerm_sql_virtual_network_rule" "vnet_rule" {
-  for_each            = { for subnet in local.allowed_subnets : subnet.name => subnet }
+  for_each            = try({ for subnet in local.allowed_subnets : subnet.name => subnet }, {})
   name                = each.key
   resource_group_name = var.resource_group_name
   server_name         = azurerm_mssql_server.sql.name