+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Breaking
 
 Changed
   * AZ-908: Bump `diagnostics-settings`
+  * AZ-908: Rework and optimize module HCL code
 
 # v7.1.0 - 2022-11-18